+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Simple list component (#84)
 - Skeleton component (#85)
 - Toggle group component (#97)
+- Truncate component (#101)
 
 ## [0.1.4] - 2024-01-25